Consolidate Mictrack protocol variants into single MictrackProtocol
Replaces separate MictrackHQProtocol and MictrackMT700Protocol with a
unified protocol class that handles all Mictrack device families on a
single configurable port. A new MictrackFrameDecoder detects the frame
format from the first byte (* for HQ, # for MT700-style, else newline)
allowing the shared MictrackProtocolDecoder to route to the appropriate
parsing path. A MictrackProtocolEncoder reads a channel attribute set by
the decoder to select the correct command encoding (HQ vs MT700).

Alarm handling is now model-aware: MT700/MT700W DEF maps to ALARM_REMOVING
(light sensor / device removal alert) while MT600 DEF maps to ALARM_POWER_CUT,
and MT700 HT is correctly treated as a mode indicator rather than a
temperature alarm. V4 heartbeat messages send the required R12 ACK and
return null. V5 and V6 extra fields are parsed via optional pattern groups
rather than fragile string manipulation.
