--- src/main/java/org/traccar/protocol/MictrackHQProtocol.java
@@ -1,61 +1,0 @@
-/*
- * Copyright 2026 Drew Taylor (Drew.Taylor@fognetx.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.traccar.protocol;
-
-import io.netty.handler.codec.string.StringDecoder;
-import io.netty.handler.codec.string.StringEncoder;
-import org.traccar.BaseProtocol;
-import org.traccar.CharacterDelimiterFrameDecoder;
-import org.traccar.PipelineBuilder;
-import org.traccar.TrackerServer;
-import org.traccar.config.Config;
-import org.traccar.model.Command;
-
-import jakarta.inject.Inject;
-
-public class MictrackHQProtocol extends BaseProtocol {
-
-    @Inject
-    public MictrackHQProtocol(Config config) {
-        setSupportedDataCommands(
-                Command.TYPE_CUSTOM,
-                Command.TYPE_POSITION_PERIODIC,
-                Command.TYPE_ENGINE_STOP,
-                Command.TYPE_ENGINE_RESUME,
-                Command.TYPE_ALARM_ARM,
-                Command.TYPE_ALARM_DISARM);
-        addServer(new TrackerServer(config, getName(), false) {
-            @Override
-            protected void addProtocolHandlers(PipelineBuilder pipeline, Config config) {
-                pipeline.addLast(new CharacterDelimiterFrameDecoder(1024, '#'));
-                pipeline.addLast(new StringEncoder());
-                pipeline.addLast(new StringDecoder());
-                pipeline.addLast(new MictrackHQProtocolEncoder(MictrackHQProtocol.this));
-                pipeline.addLast(new MictrackHQProtocolDecoder(MictrackHQProtocol.this));
-            }
-        });
-        addServer(new TrackerServer(config, getName(), true) {
-            @Override
-            protected void addProtocolHandlers(PipelineBuilder pipeline, Config config) {
-                pipeline.addLast(new StringEncoder());
-                pipeline.addLast(new StringDecoder());
-                pipeline.addLast(new MictrackHQProtocolEncoder(MictrackHQProtocol.this));
-                pipeline.addLast(new MictrackHQProtocolDecoder(MictrackHQProtocol.this));
-            }
-        });
-    }
-
-}
